refator: modularizando tela que estava na main
